Feat: Follow 화면 ui 수정
