+++ feature/follower/src/main/res/values/strings.xml
@@ -5,4 +5,5 @@
     <string name="string_not_follow">팔로잉 하고 있는 사람이 없습니다.</string>
     <string name="string_error">오류가 발생하였습니다!</string>
     <string name="string_retry">Retry</string>
+    <string name="string_edit">수정</string>
 </resources>
